Toad's scripts for provisioning.
These are a stupid set of scripts I used for roughly repeatable provisioning of
my cloud pseudo-prod instance.  It's all very stupid because we have

1. Salt which does the real provisioning.
2. Ansible which does some very basic stuff, that should be done by Salt, but
   I don't know how to do (this includes installing Salt).
3. provision.sh which is a shell script that runs both of the above as well
   directly rsyncing some of the scripts over.

provision.sh is able to parse enough of the `salt/salt/prod.sls` config to
figure out the hostname.  But this must match the hostname in the ansible
inventory.

This commit has the actual provision-with-salt step commented out, because the
salt configs aren't yet correct.